+++ AlphaFS.UnitTest/AlphaFS Volume Class/AlphaFS.Volume.QueryDosDevice.cs
@@ -21,7 +21,6 @@
 
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using System;
-using System.Collections.Generic;
 using System.Linq;
 
 namespace AlphaFS.UnitTest
@@ -32,130 +31,24 @@ namespace AlphaFS.UnitTest
 
 
       [TestMethod]
-      public void AlphaFS_Volume_QueryAllDosDevices()
+      public void AlphaFS_Volume_QueryAllDosDevices_Local_Success()
       {
          UnitTestConstants.PrintUnitTestHeader(false);
          Console.WriteLine();
 
 
-         IEnumerable<string> query = Alphaleonis.Win32.Filesystem.Volume.QueryAllDosDevices("sort").ToArray();
+         var query = Alphaleonis.Win32.Filesystem.Volume.QueryAllDosDevices().ToArray();
 
-         Console.WriteLine("Retrieved: [{0}] items.\n", query.Count());
+         Console.WriteLine("Retrieved {0} MS-DOS Device Names.\n", query.Length);
 
-
-         var cnt = 0;
-         foreach (var dosDev in query)
-         {
-            Console.WriteLine("#{0:000}\t{1}", ++cnt, dosDev);
-         }
-
-         if (cnt == 0)
-            Assert.Inconclusive("Nothing is enumerated, but it is expected.");
-
-         Assert.IsTrue(query.Any());
-      }
+         Assert.IsTrue(query.Length > 0, "No MS-DOS device names enumerated, but it is expected.");
 
 
-      [TestMethod]
-      public void AlphaFS_Volume_QueryDosDevice_Local_Success()
-      {
-         UnitTestConstants.PrintUnitTestHeader(false);
-         Console.WriteLine();
+         var deviceNameCount = 0;
 
+         foreach (var deviceName in query)
 
-         #region Filtered, UnSorted List Drives
-
-         var filter = "hard";
-         
-         IEnumerable<string> query = Alphaleonis.Win32.Filesystem.Volume.QueryDosDevice(filter).ToArray();
-
-         Console.WriteLine("\tQueryDosDevice(\"" + filter + "\")\n\tRetrieved Filtered, UnSorted list: [{0}]\n\tList .Count(): [{1}]\n", query.Any(), query.Count());
-
-         var cnt = 0;
-
-         foreach (var dosDev in query)
-            Console.WriteLine("\t\t#{0:000}\tMS-Dos Name: [{1}]", ++cnt, dosDev);
-
-         Assert.IsTrue(query.Any() && cnt > 0);
-
-         #endregion // Filtered, UnSorted List Drives
-
-
-         #region Filtered, Sorted List Volumes
-
-         filter = "volume";
-
-         cnt = 0;
-
-         query = Alphaleonis.Win32.Filesystem.Volume.QueryDosDevice(filter, "sort").ToArray();
-
-         Console.WriteLine("\n\n\tQueryDosDevice(\"" + filter + "\")\n\tRetrieved Filtered, Sorted list: [{0}]\n\tList .Count(): [{1}]\n", query.Any(), query.Count());
-
-         foreach (var dosDev in query)
-            Console.WriteLine("\t\t#{0:000}\tMS-Dos Name: [{1}]", ++cnt, dosDev);
-
-         Assert.IsTrue(query.Any() && cnt > 0);
-
-         #endregion // Filtered, Sorted List Volumes
-
-
-         #region Get from Logical Drives
-
-         Console.WriteLine("\n\n\tQueryDosDevice (from Logical Drives)\n");
-
-         cnt = 0;
-
-         foreach (var existingDriveLetter in System.IO.Directory.GetLogicalDrives())
-         {
-            foreach (var dosDev in Alphaleonis.Win32.Filesystem.Volume.QueryDosDevice(existingDriveLetter))
-            {
-               var hasLogicalDrive = !Alphaleonis.Utils.IsNullOrWhiteSpace(dosDev);
-
-               Console.WriteLine("\t\t#{0:000}\tLogical Drive [{1}] MS-Dos Name: [{2}]", ++cnt, existingDriveLetter, dosDev);
-
-               Assert.IsTrue(hasLogicalDrive);
-            }
-         }
-
-         Assert.IsTrue(cnt > 0, "No entries read.");
-
-         #endregion // Get from Logical Drives
-
-
-         #region Filtered, Sorted PhysicalDrive
-
-         filter = "PhysicalDrive";
-
-         cnt = 0;
-
-         query = Alphaleonis.Win32.Filesystem.Volume.QueryDosDevice(filter, "sort").ToArray();
-
-         Console.WriteLine("\n\n\tQueryDosDevice(\"" + filter + "\")\n\tRetrieved Filtered, Sorted list: [{0}]\n\tList .Count(): [{1}]\n", query.Any(), query.Count());
-
-         foreach (var dosDev in query)
-            Console.WriteLine("\t\t#{0:000}\tMS-Dos Name: [{1}]", ++cnt, dosDev);
-
-         Assert.IsTrue(query.Any() && cnt > 0);
-
-         #endregion // Filtered, Sorted PhysicalDrive
-
-
-         #region Filtered, Sorted CDRom
-
-         filter = "CdRom";
-
-         cnt = 0;
-
-         UnitTestConstants.StopWatcher(true);
-
-         query = Alphaleonis.Win32.Filesystem.Volume.QueryDosDevice(filter, "sort").ToArray();
-
-         Console.WriteLine("\n\n\tQueryDosDevice(\"" + filter + "\")\n\tRetrieved Filtered, Sorted list: [{0}]\n\tList .Count(): [{1}]\n", query.Any(), query.Count());
-
-         foreach (var dosDev in query)
-            Console.WriteLine("\t\t#{0:000}\tMS-Dos Name: [{1}]", ++cnt, dosDev);
-
-         #endregion // Filtered, Sorted CDRom
+            Console.WriteLine("\t#{0:000}\t{1}", ++deviceNameCount, deviceName);
       }
    }
 }